+++ R/check_input_data.R
@@ -44,7 +44,38 @@ check_input_data <- function(y, z, X){
                "observation than response and treatment vectors.",
                "Current values: ", covars_size, ", ", y_size))
   }
+
+  #TODO: convert all variables to a data.table.
+  if (!(class(y)[[1]] %in% c("numeric", "integer", "matrix"))){
+    stop("Invalid 'y' input. Please input a numeric vector.")
+  }
+
+  if (length(unique(z)) != 2){
+    stop("Invalid 'z' input. Please input a binary treatment vector.")
+  }
+
+  if (length(class(X)) == 1) {
+    if (!(class(X) %in% c("data.frame", "matrix"))) {
+      stop("Invalid 'X' input. Please input a matrix or data frame.")
+    }
+  }
+
+  if (length(class(X)) == 2) {
+    if (!(identical(class(X), c("matrix", "array")))) {
+      stop("Invalid 'X' input. Please input a matrix or data frame.")
+    }
+  }
+
+  X_classes <- apply(X, 2, class)
+
+  if (!all(X_classes %in% c("integer", "numeric"))){
+    stop(paste("Invalid 'X' input. Please input a matrix or data frame",
+               " of numeric categorical variables"))
+  }
+
   #-----------------------------------------------------------------------------
 
   invisible(y_size)
 }
+
+