+++ NEWS.md
@@ -8,11 +8,13 @@
 * improve examples and update tests for all functions
 
 ### Added
+* `check_input` function to isolate input checks.
 * `estimate_ite_aipw` function for augmented inverse propensity weighting
 * `plot.cre` generic function to plot CRE S3 object Results
 * `test-cre_functional.R` tests the functionality of the package
 
 ### Removed
+
 * `estimate_ite_blp` function
 
 ## CRE 0.0.1 (2021-10-20)